Breaking: Add Vitest support, assign mockedWindow to window for Vitest support
