+++ package.json
@@ -33,7 +33,9 @@
 		"build": "tsc -p ./tsconfig-build.json",
 		"lint": "eslint",
 		"prepare": "husky",
-		"test": "jest"
+		"test": "npm run test:jest && npm run test:vitest",
+		"test:jest": "jest",
+		"test:vitest": "vitest"
 	},
 	"commitlint": {
 		"extends": "./node_modules/conventional-changelog-evelyn/commitlint.config.js"
@@ -55,27 +57,38 @@
 		"@testing-library/jest-dom": "^6.9.1",
 		"@testing-library/react": "^16.3.2",
 		"@types/jest": "^30.0.0",
-		"@types/react": "^19.2.14",
+		"@types/react": "^19.2.17",
 		"@types/react-dom": "^19.2.3",
-		"commitlint": "^20.4.3",
+		"@vitest/coverage-v8": "^4.1.9",
+		"commitlint": "^21.0.2",
 		"conventional-changelog-evelyn": "^2.0.1",
 		"eslint": "^9.39.3",
 		"eslint-plugin-evelyn": "^12.1.0",
 		"husky": "^9.1.7",
-		"jest": "^30.2.0",
-		"jest-environment-jsdom": "^30.2.0",
-		"lint-staged": "^16.3.2",
-		"react": "^19.2.4",
-		"react-dom": "^19.2.4",
-		"react-router-dom": "^7.13.1",
-		"semantic-release": "^25.0.3",
-		"sort-package-json": "^3.6.1",
-		"ts-jest": "^29.4.6",
+		"jest": "^30.4.2",
+		"jest-environment-jsdom": "^30.4.1",
+		"lint-staged": "^17.0.7",
+		"react": "^19.2.7",
+		"react-dom": "^19.2.7",
+		"react-router-dom": "^7.18.0",
+		"semantic-release": "^25.0.5",
+		"sort-package-json": "^4.0.0",
+		"ts-jest": "^29.4.11",
 		"typescript": "^5.9.3",
+		"vitest": "^4.1.9",
 		"whatwg-fetch": "^3.6.20"
 	},
 	"peerDependencies": {
-		"@jest/globals": "^30.0.0"
+		"@jest/globals": "^30.0.0",
+		"vitest": "^4.0.0"
+	},
+	"peerDependenciesMeta": {
+		"@jest/globals": {
+			"optional": true
+		},
+		"vitest": {
+			"optional": true
+		}
 	},
 	"engines": {
 		"node": ">=20.0.0"
