+++ src/__tests__/react-router-dom.vitest.test.tsx
@@ -1,0 +1,20 @@
+import {
+	describe,
+	it,
+	expect,
+	beforeEach,
+	afterEach,
+	beforeAll,
+	afterAll,
+} from "vitest";
+import {testSuite} from "./test-suites/react-router-dom";
+
+testSuite({
+	describe,
+	it,
+	expect,
+	beforeEach,
+	afterEach,
+	beforeAll,
+	afterAll,
+});
