+++ src/__tests__/spies.jest.test.ts
@@ -1,0 +1,20 @@
+import {
+	describe,
+	it,
+	expect,
+	beforeEach,
+	afterEach,
+	beforeAll,
+	afterAll,
+} from "@jest/globals";
+import {testSuite} from "./test-suites/spies";
+
+testSuite({
+	describe,
+	it,
+	expect,
+	beforeEach,
+	afterEach,
+	beforeAll,
+	afterAll,
+});
