+++ src/hooks/__tests__/replace-history-node.jest.test.ts
@@ -1,0 +1,23 @@
+/** @jest-environment node */
+import {
+	describe,
+	it,
+	expect,
+	beforeEach,
+	afterEach,
+	beforeAll,
+	afterAll,
+} from "@jest/globals";
+import {replaceHistory} from "../replace-history.jest";
+import {testSuite} from "./test-suites/replace-history-node";
+
+testSuite({
+	describe,
+	it,
+	expect,
+	beforeEach,
+	afterEach,
+	beforeAll,
+	afterAll,
+	replaceHistory,
+});
