+++ src/hooks/__tests__/replace-history-node.vitest.test.ts
@@ -1,0 +1,23 @@
+/** @vitest-environment node */
+import {
+	describe,
+	it,
+	expect,
+	beforeEach,
+	afterEach,
+	beforeAll,
+	afterAll,
+} from "vitest";
+import {replaceHistory} from "../replace-history.vitest";
+import {testSuite} from "./test-suites/replace-history-node";
+
+testSuite({
+	describe,
+	it,
+	expect,
+	beforeEach,
+	afterEach,
+	beforeAll,
+	afterAll,
+	replaceHistory,
+});
