+++ src/hooks/__tests__/replace-location.jest.test.ts
@@ -1,0 +1,22 @@
+import {
+	describe,
+	it,
+	expect,
+	beforeEach,
+	afterEach,
+	beforeAll,
+	afterAll,
+	jest,
+} from "@jest/globals";
+import {testSuite} from "./test-suites/replace-location";
+
+testSuite({
+	describe,
+	it,
+	expect,
+	beforeEach,
+	afterEach,
+	beforeAll,
+	afterAll,
+	fn: jest.fn,
+});
