+++ src/hooks/__tests__/replace-location.vitest.test.ts
@@ -1,0 +1,22 @@
+import {
+	describe,
+	it,
+	expect,
+	beforeEach,
+	afterEach,
+	beforeAll,
+	afterAll,
+	vi,
+} from "vitest";
+import {testSuite} from "./test-suites/replace-location";
+
+testSuite({
+	describe,
+	it,
+	expect,
+	beforeEach,
+	afterEach,
+	beforeAll,
+	afterAll,
+	fn: vi.fn,
+});
