+++ src/hooks/__tests__/test-suites/replace-history-node.ts
@@ -1,0 +1,21 @@
+import type {TestRunnerDependencies} from "../../../../config/create-test-suite";
+import type {replaceHistory as replaceHistoryType} from "../../replace-history.jest";
+
+export const testSuite = ({
+	describe,
+	it,
+	expect,
+	replaceHistory,
+}: TestRunnerDependencies & {
+	replaceHistory: typeof replaceHistoryType;
+}) => {
+	describe("replaceHistory()", () => {
+		describe("when window not defined", () => {
+			it("should not change window", () => {
+				// Should already be run by jest setup, but let's run again for test verbosity and to report any errors
+				replaceHistory();
+				expect(typeof window).toBe("undefined");
+			});
+		});
+	});
+};
