+++ src/hooks/replace-history.jest.ts
@@ -1,0 +1,4 @@
+import {jest} from "@jest/globals";
+import {replaceHistory as replaceHistoryBase} from "./replace-history";
+
+export const replaceHistory = () => replaceHistoryBase(jest.spyOn);
