+++ src/hooks/replace-history.vitest.ts
@@ -1,0 +1,4 @@
+import {vi} from "vitest";
+import {replaceHistory as replaceHistoryBase} from "./replace-history";
+
+export const replaceHistory = () => replaceHistoryBase(vi.spyOn);
