+++ src/hooks/replace-location.jest.ts
@@ -1,0 +1,4 @@
+import {jest} from "@jest/globals";
+import {replaceLocation as replaceLocationBase} from "./replace-location";
+
+export const replaceLocation = () => replaceLocationBase(jest.spyOn);
