+++ src/hooks/replace-location.vitest.ts
@@ -1,0 +1,4 @@
+import {vi} from "vitest";
+import {replaceLocation as replaceLocationBase} from "./replace-location";
+
+export const replaceLocation = () => replaceLocationBase(vi.spyOn);
