+++ src/hooks/vitest.ts
@@ -1,0 +1,3 @@
+export * from "./replace-location.vitest";
+export * from "./replace-history.vitest";
+export * from "./reset";
