+++ src/utils/__tests__/location-mock-relative.vitest.test.ts
@@ -1,0 +1,20 @@
+import {
+	describe,
+	it,
+	expect,
+	beforeEach,
+	afterEach,
+	beforeAll,
+	afterAll,
+} from "vitest";
+import {testSuite} from "./test-suites/location-mock-relative";
+
+testSuite({
+	describe,
+	it,
+	expect,
+	beforeEach,
+	afterEach,
+	beforeAll,
+	afterAll,
+});
